+++ draw-it/src/ui.rs
@@ -203,7 +203,7 @@ impl Ui {
         let mut uvs = vec![];
         let mut to = 0;
         for draw_list in draw_data.draw_lists() {
-            indices.extend(draw_list.idx_buffer().iter().map(|i| *i as u32 + to));
+            indices.extend(draw_list.idx_buffer().iter().map(|i| *i + to));
             for vert in draw_list.vtx_buffer() {
                 let vertex =
                     Vector3::new(vert.pos[0] - half_width, -vert.pos[1] + half_height, 1.0);
@@ -214,7 +214,7 @@ impl Ui {
                 colors.push(color);
                 normals.push(Vector3::BACKWARD);
             }
-            to = vertices.len() as u32;
+            to = vertices.len() as u16;
         }
 
         // update mesh
